requirements.txt
fastapi
uvicorn[standard]
openai
faiss-cpu
numpy
python-dotenv
pydantic
sentence-transformers
torch
scikit-learn
requests
